Move move context code into new `._context` mod
